application controller: revert calculation of overhead releases in cleanUpReleasesForApplication()
This change has erroneously removed the incumbent release as soon the
contender release is created. Additionally a safe-guard was introduced
to protect cases where the user might have a revision history limit
smaller than 1.

The respective test was modified to honor that situation, expecting that
the oldest release to be removed, meaning that effectively the releases
present in the application's `.status.history` is the contender plus
`.spec.revisionHistoryLimit` past releases.